requirements.txt
Django==2.2.7
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
flake8
autopep8
behave-django